Compile but doesn't show anything 13/08

+++ CommonCore/fdf/draw_manipulation.c
@@ -6,7 +6,7 @@
 /*   By: emedeiro <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/13 14:32:39 by emedeiro          #+#    #+#             */
-/*   Updated: 2024/08/13 20:29:54 by emedeiro         ###   ########.fr       */
+/*   Updated: 2024/08/13 23:29:37 by emedeiro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ void bresenham_algorithm(float x_a, float y_a, float x_b, float y_b, t_fdf *win)
     y = y_a;
     while ((int)(x - x_b) || (int)(y - y_b))
     {
-        mlx_pixel_put(win->mlx, win->win, (int)x, (int)y, 0xFFFFFF);
+        mlx_pixel_put(win->mlx, win->win, x, y, 0xFFFFFF);
         x += Pa;
         y += Pb;
     }
